Multiple Pots Alpha
Refined a bit of the client to be able to show multiple servo readings.
Changed the broadcast to a JSON’d object with servoNum and pot reading.

It’s not working. Well, for some reason the receiveSerialData is
receiving random readings. Sometimes it’s all 3 pots, sometimes 2, and
sometimes 1. But the calls seem to be made in order. Maybe I will look
into synchronizing the calls to receiveSerialData or implementing a
locking mechanism.